# src/main/java/com/benz/tutorial/jhipster/gateway/web/api/MoviesApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.2.3).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.benz.tutorial.jhipster.gateway.web.api;

import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import java.util.List;
import javax.validation.constraints.*;
import javax.validation.Valid;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2018-02-15T16:15:07.332-08:00")

@Api(value = "movies", description = "the movies API")
public interface MoviesApi {

	@ApiOperation(value = "Get movie based on id", notes = "", response = Void.class, tags = { "movieid", })
	@ApiResponses(value = { @ApiResponse(code = 200, message = "Successful resoponse", response = Void.class) })

	@RequestMapping(value = "/movies/{id}", method = RequestMethod.GET)
	public ResponseEntity<Void> findMovie(@ApiParam(value = "", required = true) @PathVariable("id") String id);

	@ApiOperation(value = "Gets all the movies. Even helps do search by title", notes = "", response = Void.class, tags = {
			"movies", })
	@ApiResponses(value = { @ApiResponse(code = 200, message = "Successful response", response = Void.class) })

	@RequestMapping(value = "/movies", produces = { "application/xml", "application/json" }, consumes = {
			"application/xml", "application/json" }, method = RequestMethod.GET)
	public ResponseEntity<Void> getAllMovies();

}